bootcamp-bancopan-aulas/java/JUnitClass/src/main/java/demo/Calculadora.java
package demo;

public class Calculadora {

    private double resultado = 0;

    public double aumentar(int n1, int n2) {
        resultado = n1 + n2;
        return resultado;
    }

    public double diminuir(int n1, int n2) {
        resultado = n1 - n2;
        return resultado;
    }

    public double multiplicar(int n1, int n2) {
        resultado = n1 * n2;
        return resultado;
    }

    public double dividir(int n1, int n2) {
        resultado = n1 / n2;
        return resultado;
    }

    public double duplica (int n1) {
        resultado = n1 * 2;
        return resultado;
    }

}
